Textarea: Added textarea component.
Texarea component of the framework. Extended from Field component to be consistent with rest of the framework. However it uses `goog.ui.Textarea` in the background. This is because goog handles resizing the textarea to fit its content pretty good. This might be a disadvantage for the filesize. I need to consider it in the future.