View: Added render method with renderTo option.
Renders element into another View instance or a DOM element. Be aware that this method only appends the element of this class into another element. If the parent element is not in DOM, this method will not append the parent into DOM. Also you can use this method with `renderTo` option. If you pass `renderTo` as a View instance or DOM element, View will call render method by default.