Store: `set` method now validates the given value before set.
Sets a new key value pair into store. It will also make a validation before setting the new value. If you set a validation rule in options.validations for that property, when validation failed, the property won't be updated. Store::set will return no and emit a 'PropertyRejected' event otherwise it will return true and it will also emit 'ProperySet'. Both events will include key and value. If the key and value passed to set does not related with a validation rule, it will directly set to store. This makes validation optional.