Utils: Added `spark.utils.parseTemplateTags`
Replaces mustache like template tags with the value in the given data. Currently only parses variable tags like `{{name}}`. I am planning to make it more powerful with upcoming releases. Tags are not whitespace sensitive. So `{{ name }}` and `{{name}}` and `{{   name   }}` are the same. Here is a usage example. Replaced values are html escaped to avoid possible XSS attempt.